backend/requirements.txt
# ProjectMeats Backend Requirements
# Django REST Framework backend for PowerApps migration

# Core Django and DRF
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.3

# Multi-tenancy
django-tenants==3.5.0

# Database
psycopg[binary]==3.2.9  # PostgreSQL adapter (modern replacement for psycopg2-binary, better Python 3.13+ support)
dj-database-url==2.1.0

# API Documentation
drf-spectacular==0.27.0

# Development and Testing
django-extensions==3.2.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code Quality
black==23.11.0
flake8==6.1.0
isort==5.12.0
pre-commit==3.5.0

# Production (optional for development)
gunicorn==21.2.0
whitenoise==6.6.0

# Utilities
python-decouple==3.8
Pillow>=10.2.0  # For ImageField support if needed - 10.2.0+ supports Python 3.13
psutil>=5.9.0  # For system health monitoring

# HTTP requests
requests>=2.31.0  # For GitHub API integration in bug reports